move hicdata into server/routes/
