+++ server/src/bam.js
@@ -5,7 +5,6 @@ import serverconfig from './serverconfig'
 import { spawn } from 'child_process'
 import { Readable, pipeline } from 'stream'
 import { createCanvas } from 'canvas'
-import { readline } from 'readline'
 import * as bamcommon from './bam.common'
 import { run_rust } from '@sjcrh/proteinpaint-rust'
 import crypto from 'crypto'
